Refactor
Move token meta information to lexer.rs
May combine with tokens at a later time
